+++ tests/test_stringio.rs
@@ -26,8 +26,8 @@ mod test_stream_stringio {
     #[test]
     fn test_out() {
         let sout = StringOut::default();
-        let res = sout
-            .lock()
+        #[rustfmt::skip]
+        let res = sout.lock()
             .write_fmt(format_args!("{}\nACBDE\nefgh\n", 1234));
         assert!(res.is_ok());
         assert_eq!(sout.lock().buffer_str(), "1234\nACBDE\nefgh\n");
@@ -35,8 +35,8 @@ mod test_stream_stringio {
     #[test]
     fn test_err() {
         let serr = StringErr::default();
-        let res = serr
-            .lock()
+        #[rustfmt::skip]
+        let res = serr.lock()
             .write_fmt(format_args!("{}\nACBDE\nefgh\n", 1234));
         assert!(res.is_ok());
         assert_eq!(serr.lock().buffer_str(), "1234\nACBDE\nefgh\n");
@@ -44,35 +44,32 @@ mod test_stream_stringio {
 }
 
 mod test_stream_ioe_stringio {
-    use runnel::medium::stringio::*;
     use runnel::*;
     use std::io::BufRead;
     use std::io::Write;
 
     #[test]
     fn test_ioe() {
-        let sioe = StreamIoe {
-            pin: Box::new(StringIn::with_str("ABCDE\nefgh\n")),
-            pout: Box::new(StringOut::default()),
-            perr: Box::new(StringErr::default()),
-        };
+        let sioe = RunnelIoeBuilder::new()
+            .fill_stringio_wit_str("ABCDE\nefgh\n")
+            .build();
         {
-            let mut lines_iter = sioe.pin.lock().lines().map(|l| l.unwrap());
+            let mut lines_iter = sioe.pin().lock().lines().map(|l| l.unwrap());
             assert_eq!(lines_iter.next(), Some(String::from("ABCDE")));
             assert_eq!(lines_iter.next(), Some(String::from("efgh")));
             assert_eq!(lines_iter.next(), None);
             //
             #[rustfmt::skip]
-            let res = sioe.pout.lock()
+            let res = sioe.pout().lock()
                 .write_fmt(format_args!("{}\nACBDE\nefgh\n", 1234));
             assert!(res.is_ok());
             //
             #[rustfmt::skip]
-            let res = sioe.perr.lock()
+            let res = sioe.perr().lock()
                 .write_fmt(format_args!("{}\nACBDE\nefgh\n", 1234));
             assert!(res.is_ok());
         }
-        assert_eq!(sioe.pout.lock().buffer_str(), "1234\nACBDE\nefgh\n");
-        assert_eq!(sioe.perr.lock().buffer_str(), "1234\nACBDE\nefgh\n");
+        assert_eq!(sioe.pout().lock().buffer_str(), "1234\nACBDE\nefgh\n");
+        assert_eq!(sioe.perr().lock().buffer_str(), "1234\nACBDE\nefgh\n");
     }
 }